floats changed to double

+++ TrajectoryControllerNoOrientation.cpp
@@ -27,7 +27,7 @@ noOrientation::noOrientation ()
 
 
 	void 
-noOrientation::setConstants(float l1_val, float K0_val, float R_val, float r_val, double ul, double ll)
+noOrientation::setConstants(double l1_val, double K0_val, double R_val, double r_val, double ul, double ll)
 {
 	l1 = l1_val;
 	K0 = K0_val;
@@ -40,8 +40,8 @@ noOrientation::setConstants(float l1_val, float K0_val, float R_val, float r_val
 } 
 
 
-	float
-noOrientation::k (float theta_e )
+	double
+noOrientation::k (double theta_e )
 {
 	// k(d,theta_e) = 0 * cos(theta_e)
 	return K0 * cos(theta_e);
@@ -49,7 +49,7 @@ noOrientation::k (float theta_e )
 
 
         Eigen::Vector2d	
-noOrientation::update (float u1, float d, float theta_e )
+noOrientation::update (double u1, double d, double theta_e )
 {
 	double u2 = (-u1*tan(theta_e) / l1) - ( (u1 * k(theta_e) * d) / cos(theta_e));
 
@@ -61,7 +61,7 @@ noOrientation::update (float u1, float d, float theta_e )
 
 
         double	
-noOrientation::limit ( float val )
+noOrientation::limit ( double val )
 {
 	if (val > u_limit)
 	    return u_limit;
